:truck::lipstick: Create the newRecipes component and move the form inside it. Add style to the form and make it responsive.
