+++ client/src/components/newRecipes/NewRecipes.css
@@ -1,0 +1,331 @@
+.newRecipes {
+  width: 320px;
+  height: 100%;
+  padding: 0 15px 0 0;
+
+  h2 {
+    text-align: center;
+    font-size: 1.5rem;
+    font-weight: 200;
+  }
+
+  h3 {
+    font-size: 1.2rem;
+    font-weight: 200;
+    margin: 8px 0 0 0;
+  }
+
+  input[type="text"] {
+    width: 100%;
+    height: 30px;
+    padding: 6px 8px;
+    border: none;
+    border-radius: 8px;
+    font-size: 0.8rem;
+  }
+
+  label[for="image"] > legend {
+    height: 30px;
+    font-size: 0.8rem;
+    padding: 4px 8px;
+    background-color: var(--text-light);
+    color: #666;
+    cursor: pointer;
+    border-radius: 8px;
+  }
+
+  input[type="file"] {
+    display: none;
+  }
+
+  img:not([src]),
+  img[src=""] {
+    display: none;
+  }
+  img {
+    max-width: 150px;
+    max-height: 150px;
+    border-radius: 8px;
+    padding: 8px 0 0 0;
+  }
+
+  select[name="guestNumber"] {
+    width: 100%;
+    height: 30px;
+    padding: 6px 8px;
+    border: none;
+    border-radius: 8px;
+    font-size: 0.8rem;
+    cursor: pointer;
+    appearance: none;
+    background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 20 20'%3e%3cpath stroke='%236b7280' stroke-linecap='round' stroke-linejoin='round' stroke-width='1.5' d='M6 8l4 4 4-4'/%3e%3c/svg%3e");
+    background-position: right 10px center;
+    background-repeat: no-repeat;
+    background-size: 16px;
+  }
+
+  .MuiAutocomplete-root .MuiOutlinedInput-root {
+    background-color: var(--text-light);
+    height: 30px;
+    border-radius: 8px;
+    margin: 0 0 8px 0;
+  }
+
+  .css-113d811-MuiFormLabel-root-MuiInputLabel-root {
+    font-size: 0.6rem;
+    top: 3px;
+  }
+
+  .css-19qnlrw-MuiFormLabel-root-MuiInputLabel-root {
+    font-size: 0.8rem;
+    top: -11px;
+  }
+
+  .css-1x7n7v0-MuiButtonBase-root-MuiIconButton-root-MuiAutocomplete-clearIndicator {
+    padding: 0;
+  }
+
+  .ingredient-inputs {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    gap: 13px;
+    margin: 16px 0;
+  }
+
+  fieldset[name="ingredients"] p {
+    flex-direction: column;
+    gap: 10px;
+  }
+
+  input[type="number"] {
+    padding: 6px 8px;
+    height: 30px;
+    border: none;
+    border-radius: 8px;
+    font-size: 0.8rem;
+    margin: 4px 0 0 0;
+  }
+
+  select[name="unit"] {
+    height: 30px;
+    padding: 6px 8px;
+    margin: 4px 0 0 0;
+    border: none;
+    border-radius: 8px;
+    font-size: 0.8rem;
+    appearance: none;
+    cursor: pointer;
+    background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 20 20'%3e%3cpath stroke='%236b7280' stroke-linecap='round' stroke-linejoin='round' stroke-width='1.5' d='M6 8l4 4 4-4'/%3e%3c/svg%3e");
+    background-position: right 0px center;
+    background-repeat: no-repeat;
+    background-size: 16px;
+  }
+
+  button {
+    padding: 6px 16px;
+    border: none;
+    border-radius: 8px;
+    font-size: 0.8rem;
+    cursor: pointer;
+    background-color: var(--text-light);
+    color: var(--light-primary);
+  }
+
+  button:hover {
+    background-color: rgb(226, 226, 226);
+  }
+
+  .remove-ingredient-btn,
+  .remove-step-btn {
+    background-color: var(--light-secondary) !important;
+    color: white !important;
+    padding: 6px 12px !important;
+    font-size: 0.6rem;
+  }
+
+  .remove-ingredient-btn:hover,
+  .remove-step-btn:hover {
+    background-color: #f07c09 !important;
+  }
+
+  fieldset {
+    border: none;
+  }
+
+  fieldset[name="ingredients"] {
+    background-color: rgba(255, 255, 255, 0.1);
+    border-radius: 8px;
+    padding: 8px 16px;
+    margin: 0 0 16px 0;
+  }
+
+  fieldset[name="ingredients"] > legend {
+    color: white;
+    font-size: 1rem;
+    font-weight: 200;
+  }
+
+  fieldset[name="ingredients"] p {
+    background-color: var(--text-light);
+    color: var(--text-dark);
+    padding: 8px 16px;
+    margin: 0 0 8px 0;
+    border-radius: 8px;
+    display: flex;
+    font-size: 0.8rem;
+    align-items: center;
+  }
+
+  fieldset:not([name="ingredients"]) legend {
+    color: white;
+    font-size: 1.2rem;
+    font-weight: 200;
+    margin: 0 0 8px 0;
+    display: flex;
+    align-items: center;
+    gap: 8px;
+  }
+
+  textarea {
+    width: 100%;
+    min-height: 60px;
+    padding: 16px;
+    border: none;
+    border-radius: 8px;
+    font-size: 0.8rem;
+    font-family: inherit;
+    resize: vertical;
+  }
+  .buttons {
+    display: flex;
+    justify-content: space-between;
+    padding: 8px 0;
+  }
+}
+
+@media screen and (min-width: 900px) {
+  .newRecipes {
+    width: 560px;
+    height: 700px;
+    overflow-y: scroll;
+    scrollbar-width: none;
+
+    h2 {
+      font-size: 2rem;
+    }
+
+    h3 {
+      font-size: 1.5rem;
+    }
+
+    input[type="text"] {
+      height: 48px;
+      padding: 12px 16px;
+      border-radius: 10px;
+      font-size: 1rem;
+    }
+
+    label[for="image"] > legend {
+      height: 48px;
+      padding: 10px 16px;
+      border-radius: 10px;
+      font-size: 1rem;
+    }
+
+    img {
+      max-width: 300px;
+      max-height: 200px;
+      border-radius: 15px;
+    }
+
+    select[name="guestNumber"] {
+      height: 48px;
+      padding: 12px 16px;
+      border-radius: 10px;
+      font-size: 1rem;
+      background-position: right 16px center;
+    }
+
+    .MuiAutocomplete-root .MuiOutlinedInput-root {
+      height: 48px;
+      border-radius: 10px;
+    }
+
+    .css-113d811-MuiFormLabel-root-MuiInputLabel-root {
+      font-size: initial;
+      top: -1px;
+    }
+
+    .css-19qnlrw-MuiFormLabel-root-MuiInputLabel-root {
+      font-size: initial;
+      top: -4px;
+    }
+
+    .ingredient-inputs {
+      display: flex;
+      flex-wrap: nowrap;
+      align-items: center;
+      justify-content: space-between;
+    }
+
+    input[type="number"] {
+      height: 48px;
+      padding: 12px 16px;
+      border-radius: 10px;
+      font-size: 1rem;
+      margin: 0;
+    }
+
+    select[name="unit"] {
+      height: 48px;
+      width: 40%;
+      padding: 12px 16px;
+      border-radius: 10px;
+      font-size: 1rem;
+      margin: 0;
+    }
+
+    button {
+      padding: 12px 24px;
+      border-radius: 10px;
+      font-size: 1rem;
+    }
+
+    .remove-ingredient-btn,
+    .remove-step-btn {
+      padding: 8px 16px !important;
+      font-size: 1rem;
+    }
+
+    fieldset[name="ingredients"] {
+      border-radius: 15px;
+      padding: 16px 24px;
+    }
+
+    fieldset[name="ingredients"] legend {
+      font-size: 1.5rem;
+    }
+
+    fieldset[name="ingredients"] p {
+      padding: 8px 16px;
+      margin: 0 0 16px 0;
+      border-radius: 10px;
+      font-size: 1rem;
+      flex-direction: row;
+      justify-content: space-between;
+    }
+
+    fieldset:not([name="ingredients"]) legend {
+      font-size: 1.5rem;
+      margin: 16px;
+      gap: 15px;
+    }
+
+    textarea {
+      min-height: 80px;
+      border-radius: 15px;
+      font-size: 1rem;
+    }
+  }
+}
