+++ client/src/pages/kitchen/myrecipes/MyRecipes.css
@@ -1,30 +1,46 @@
 .myRecipes {
+  width: 100%;
+  height: 100%;
   display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  grid-template-rows: repeat(2, 1fr);
+  grid-template-columns: 1fr;
+  grid-template-rows: repeat(3, 1fr);
   place-items: center;
 
   > article {
     background-color: var(--light-primary);
     border-radius: 16px;
-    padding: 32px;
+    padding: 16px;
     color: white;
   }
 
   :nth-child(1) {
-    grid-area: 1 / 1 / 2 / 2;
-  }
-  :nth-child(2) {
     grid-area: 2 / 1 / 3 / 2;
   }
-  :nth-child(3) {
-    grid-area: 1 / 2 / 3 / 3;
+  :nth-child(2) {
+    grid-area: 3 / 1 / 4 / 2;
   }
-  img {
-    max-width: 150px;
+  :nth-child(3) {
+    grid-area: 1 / 1 / 2 / 2;
+  }
+}
+
+@media (min-width: 900px) {
+  .myRecipes {
+    grid-template-columns: repeat(2, 1fr);
+    grid-template-rows: repeat(2, 1fr);
+
+    > article {
+    padding: 24px;
   }
 
-  input {
-    border: solid red;
+    :nth-child(1) {
+      grid-area: 1 / 1 / 2 / 2;
+    }
+    :nth-child(2) {
+      grid-area: 2 / 1 / 3 / 2;
+    }
+    :nth-child(3) {
+      grid-area: 1 / 2 / 3 / 3;
+    }
   }
 }
